Dedupe some PSP `rtop`
