+++ config/splat.us.strtop.yaml
@@ -50,7 +50,8 @@ segments:
       - [0x670, .data, e_lock_camera]
       - [0x6B8, .data, e_breakable]
       - [0x7C8, .data, d_prize_drops]
-      - [0x80C, .data, 20B48]
+      - [0x80C, .data, e_secret_stairs]
+      - [0x8BC, .data, 21108]
       - [0x910, .data, st_update]
       - [0x95C, .data, collision]
       - [0xD1C, .data, e_red_door]
@@ -71,7 +72,8 @@ segments:
       - [0x1DEA4, .data, gen/sprites]
       - [0x1FC14, .rodata, e_richter_fight_trigger]
       - [0x1FC2C, .rodata, e_breakable]
-      - [0x1FC4C, .rodata, 20B48]
+      - [0x1FC4C, .rodata, e_secret_stairs]
+      - [0x1FC78, .rodata, 21108]
       - [0x1FC8C, .rodata, e_red_door]
       - [0x1FCA4, .rodata, e_collect]
       - [0x1FD78, .rodata, e_stage_name]
@@ -82,7 +84,8 @@ segments:
       - [0x1FEDC, c, e_lock_camera]
       - [0x203A8, c, e_richter_fight_trigger]
       - [0x204E8, c, e_breakable]
-      - [0x20B48, c]
+      - [0x20B48, c, e_secret_stairs]
+      - [0x21108, c, 21108]
       - [0x224F4, c, st_update]
       - [0x22928, c, collision]
       - [0x240E4, c, create_entity]
@@ -102,6 +105,6 @@ segments:
       - [0x2FFBC, c, e_clock_tower]
       - [0x30460, .bss, create_entity]
       - [0x30470, bss]
-      - [0x310D8, .bss, 20B48]
+      - [0x310D8, .bss, 21108]
       - [0x310DC, .bss, e_collect]
   - [0x3111C]
