+++ config/symbols.pspeu.strtop.txt
@@ -1,13 +1,10 @@
-CreateEntityFromCurrentEntity = 0x09238598;
 CreateEntityFromEntity = 0x09238620;
 EntitySkullLord = 0x092386A8;
 EntitySkullLordOutline = 0x092391B0;
 EntitySkullLordEye = 0x092395D0;
 EntitySkullLordPieces = 0x092399A0;
 EntityStageNamePopup = 0x09239B20;
-EntityIsNearPlayer = 0x0923A9A0;
 EntityRedDoor = 0x0923AA88;
-Random = 0x0923BCA8;
 EntityTombstone = 0x0923C1E0;
 EntitySoulStealOrb = 0x0923CF28;
 EntityEnemyBlood = 0x0923D488;
@@ -29,14 +26,9 @@ EntityYorickSkull = 0x092453C0;
 EntitySkeletonPieces = 0x09245880;
 DestroyEntity = 0x092459C0;
 AnimateEntity = 0x09245AE0;
-GetSideToPlayer = 0x09245D18;
 MoveEntity = 0x09245DA0;
-FallEntity = 0x09245DF0;
-UnkCollisionFunc3 = 0x09245E28;
-UnkCollisionFunc2 = 0x092460E0;
 AllocEntity = 0x092463C0;
 SetStep = 0x09246650;
-SetSubStep = 0x092466A0;
 InitializeEntity = 0x092466E0;
 EntityDummy = 0x09246888;
 GetPlayerCollisionWith = 0x09246A50;
@@ -53,7 +45,6 @@ EntitySecretStairs = 0x0924A168;
 EntityLionLamp = 0x0924A718;
 func_us_801A21F8 = 0x0924A8A0;
 EntityBackgroundBlock = 0x0924AB70;
-PlayerIsWithinHitbox = 0x0924AC80;
 EntityLockCamera = 0x0924AD80;
 EntityTriggerBeforeCastleWarp = 0x0924B400;
 EntityClockTower3D = 0x0924B518;
@@ -62,12 +53,17 @@ spriteBanks = 0x0924E960;
 rooms_layers = 0x0924EA90;
 cluts = 0x0924EB40;
 gfxBanks = 0x0924EBB0;
+g_RedDoorTiles = 0x0924EBF8;
 EntityUpdates = 0x0924EC28;
-RBO5EInitCommon = 0x0924ED40;
+g_EInitSpawner = 0x0924ED00;
+g_EInitLockCamera = 0x0924ED30;
+g_EInitCommon = 0x0924ED40;
+g_EInitRTOPCommon = 0x0924ED60;
 g_EInitTombstone = 0x0924ED70;
 g_EInitYorick = 0x0924ED80;
-g_EInitSkull = 0x0924ED90;
+g_EInitYorickSkull = 0x0924ED90;
 g_EInitSkullLord = 0x0924EDA0;
 rooms = 0x0924EDC0;
 g_StageOverlay = 0x09267500;
 BackgroundBlockInit = 0x09267668;
+D_psp_E_STAIR_SEGMENT = 0x09267D18;
