--- src/st/rtop_psp/e_clouds.c
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../rtop/rtop.h"
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/e_clouds", EntityClouds);
