--- src/st/rtop_psp/e_lock_camera.c
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../rtop/rtop.h"
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/e_lock_camera", PlayerIsWithinHitbox);
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/e_lock_camera", EntityLockCamera);
