--- src/st/rtop_psp/e_red_door.c
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../rtop/rtop.h"
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/e_red_door", EntityIsNearPlayer);
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/e_red_door", EntityRedDoor);
