+++ src/st/rtop_psp/unk_11508.c
@@ -4,15 +4,3 @@
 INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_11508", EntityTriangleElevator);
 
 INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_11508", func_us_801A1940);
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_11508", EntityStairSwitch);
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_11508", func_us_801A0BF4);
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_11508", EntitySecretStairs);
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_11508", EntityLionLamp);
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_11508", func_us_801A21F8);
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_11508", EntityBackgroundBlock);
