+++ src/st/rtop_psp/unk_13098.c
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "../rtop/rtop.h"
+
+INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_13098", EntityLionLamp);
+
+INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_13098", func_us_801A21F8);
+
+INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_13098", EntityBackgroundBlock);
