--- src/st/rtop_psp/unk_D660.c
@@ -1,8 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../rtop/rtop.h"
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_D660", EntityYorick);
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_D660", EntityYorickSkull);
-
-INCLUDE_ASM("st/rtop_psp/nonmatchings/rtop_psp/unk_D660", EntitySkeletonPieces);
